Create AboveAverage.java
Here's the context behind the code:

It is said that 90% of frosh expect to be above average in their class. You are to provide a reality check.

The first line of standard input contains an integer 1 <= C <= 50, the number of test cases. C data sets follow. Each data set begins with an integer, N, the number of people in the class (1 <= N <= 1000). N integers follow, separated by spaces or newlines, each giving the final grade (an integer between 0 and 100) of a student in the class.

For each case you are to output a line giving the percentage of students whose grade is above average, rounded to exactly 3 decimal places.